Abstract click outside logic to useClickOutside
